requirements.txt
python-telegram-bot==20.7
httpx==0.27.0
nest_asyncio
